# lepiossa-golang-2022-backend-dini-surya/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/laptop": {
            "get": {
                "description": "Get a list of Laptops.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laptop"
                ],
                "summary": "Get all laptops.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Laptop"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new Laptop.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laptop"
                ],
                "summary": "Create New Laptop.",
                "parameters": [
                    {
                        "description": "the body to create a new laptop",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.laptopInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Laptop"
                        }
                    }
                }
            }
        },
        "/laptop/{id}": {
            "get": {
                "description": "Get a Laptop by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laptop"
                ],
                "summary": "Get Laptop.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "laptop id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Laptop"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a laptop by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laptop"
                ],
                "summary": "Delete one laptop.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "laptop id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update laptop by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laptop"
                ],
                "summary": "Update Laptop.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "laptop id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update an laptop",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.laptopInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Laptop"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.laptopInput": {
            "type": "object",
            "properties": {
                "brandID": {
                    "type": "integer"
                },
                "detail": {
                    "type": "string"
                },
                "grafis": {
                    "type": "string"
                },
                "hardisk": {
                    "type": "string"
                },
                "harga": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "layar": {
                    "type": "string"
                },
                "namaLaptop": {
                    "type": "string"
                },
                "osID": {
                    "type": "integer"
                },
                "prosesor": {
                    "type": "string"
                },
                "ram": {
                    "type": "string"
                },
                "ratingID": {
                    "type": "integer"
                }
            }
        },
...
        "models.Laptop": {
            "type": "object",
            "properties": {
                "brandID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "grafis": {
                    "type": "string"
                },
                "hardisk": {
                    "type": "string"
                },
                "harga": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "layar": {
                    "type": "string"
                },
                "namaLaptop": {
                    "type": "string"
                },
                "osID": {
                    "type": "integer"
                },
                "prosesor": {
                    "type": "string"
                },
                "ram": {
                    "type": "string"
                },
                "ratingID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}